SEO Update: Changed every title tag
Changed the title tags in every html file to be more simplified
as well as contain Land Resource Design Group for SEO reasons.

I will be using LRDG in the meta description so we can grab it there.
I will have to implement LRDG in all the image alt messages. This is
important because we can drive traffic to the site from having all
these unique images we are using.